requirements.txt
django==4.2.0
djangorestframework==3.14.0
